# public/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/files/upload":{"post":{"operationId":"FilesLocalController_uploadFile","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponseDto"}}}}},"tags":["Files"],"security":[{"bearer":[]}]}},"/api/account":{"post":{"operationId":"AccountController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AccountController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/account/{id}":{"patch":{"operationId":"AccountController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AccountController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}},"tags":["Auth"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/":{"get":{"operationId":"HomeController_appInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["Home"]}},"/api/v1/group-service":{"post":{"operationId":"GroupServiceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupServiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Group Service"],"security":[{"bearer":[]}]},"get":{"operationId":"GroupServiceController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Group Service"],"security":[{"bearer":[]}]}},"/api/v1/group-service/{id}":{"put":{"operationId":"GroupServiceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["Group Service"],"security":[{"bearer":[]}]},"delete":{"operationId":"GroupServiceController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Group Service"],"security":[{"bearer":[]}]}},"/api/v1/service":{"post":{"operationId":"ServiceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Service"],"security":[{"bearer":[]}]},"get":{"operationId":"ServiceController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Service"],"security":[{"bearer":[]}]}},"/api/v1/service/{id}":{"get":{"operationId":"ServiceController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Service"],"security":[{"bearer":[]}]},"put":{"operationId":"ServiceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["Service"],"security":[{"bearer":[]}]}},"/api/v1/material":{"post":{"operationId":"MaterialController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMaterialDto"}}}},"responses":{"201":{"description":""}},"tags":["Material"],"security":[{"bearer":[]}]},"get":{"operationId":"MaterialController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Material"],"security":[{"bearer":[]}]}},"/api/v1/material/{id}":{"get":{"operationId":"MaterialController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Material"],"security":[{"bearer":[]}]},"put":{"operationId":"MaterialController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMaterialDto"}}}},"responses":{"200":{"description":""}},"tags":["Material"],"security":[{"bearer":[]}]},"delete":{"operationId":"MaterialController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Material"],"security":[{"bearer":[]}]}},"/api/v1/user-car":{"post":{"operationId":"UserCarController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserCarDto"}}}},"responses":{"201":{"description":""}},"tags":["User Car"],"security":[{"bearer":[]}]},"get":{"operationId":"UserCarController_findByUser","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User Car"],"security":[{"bearer":[]}]}},"/api/v1/user-car/{id}":{"get":{"operationId":"UserCarController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["User Car"],"security":[{"bearer":[]}]},"put":{"operationId":"UserCarController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserCarDto"}}}},"responses":{"200":{"description":""}},"tags":["User Car"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserCarController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User Car"],"security":[{"bearer":[]}]}},"/api/v1/history-car":{"post":{"operationId":"HistoryCarController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHistoryCarDto"}}}},"responses":{"201":{"description":""}},"tags":["History Car"],"security":[{"bearer":[]}]},"get":{"operationId":"HistoryCarController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["History Car"],"security":[{"bearer":[]}]},"put":{"operationId":"HistoryCarController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateHistoryCarDto"}}}},"responses":{"200":{"description":""}},"tags":["History Car"],"security":[{"bearer":[]}]}},"/api/v1/history-car/user":{"get":{"operationId":"HistoryCarController_findByUser","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["History Car"],"security":[{"bearer":[]}]}}},"info":{"title":"API","description":"API docs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"FileType":{"type":"object","properties":{"id":{"type":"string","example":"cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae"},"path":{"type":"string","example":"https://example.com/path/to/file.jpg"}},"required":["id","path"]},"FileResponseDto":{"type":"object","properties":{"file":{"$ref":"#/components/schemas/FileType"}},"required":["file"]},"CreateAccountDto":{"type":"object","properties":{"phone":{"type":"string"},"password":{"type":"string"}},"required":["phone","password"]},"UpdateAccountDto":{"type":"object","properties":{"phone":{"type":"string"},"password":{"type":"string"}}},"AuthLoginDto":{"type":"object","properties":{"phone":{"type":"string","example":"034777578421"},"password":{"type":"string"}},"required":["phone","password"]},"AuthRegisterDto":{"type":"object","properties":{"phone":{"type":"string"},"email":{"type":"string","example":"test1@example.com"},"password":{"type":"string"},"name":{"type":"string","example":"Minkey De Poet"},"dob":{"type":"string"}},"required":["phone","email","password","name","dob"]},"Account":{"type":"object","properties":{"id":{"type":"number"},"phone":{"type":"string"},"password":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","phone","password","createdAt","updatedAt","deletedAt"]},"CreateGroupServiceDto":{"type":"object","properties":{"key":{"type":"string"},"value":{"type":"string"},"status":{"type":"string"}},"required":["key","value","status"]},"UpdateGroupServiceDto":{"type":"object","properties":{"key":{"type":"string"},"value":{"type":"string"},"status":{"type":"string"}}},"CreateServiceDto":{"type":"object","properties":{"groupServiceCode":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}},"required":["groupServiceCode","name","description"]},"UpdateServiceDto":{"type":"object","properties":{"groupServiceCode":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}}},"CreateMaterialDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"importPrice":{"type":"string"},"exportPrice":{"type":"string"},"importQuantity":{"type":"number"},"exportQuantity":{"type":"number"},"groupServiceCode":{"type":"string"}},"required":["name","description","importPrice","exportPrice","importQuantity","exportQuantity","groupServiceCode"]},"UpdateMaterialDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"importPrice":{"type":"string"},"exportPrice":{"type":"string"},"importQuantity":{"type":"number"},"exportQuantity":{"type":"number"},"groupServiceCode":{"type":"string"}}},"CreateUserCarDto":{"type":"object","properties":{"mode":{"type":"string"},"plateNumber":{"type":"string"},"color":{"type":"string"}},"required":["mode","plateNumber","color"]},"UpdateUserCarDto":{"type":"object","properties":{"mode":{"type":"string"},"plateNumber":{"type":"string"},"color":{"type":"string"}}},"CreateHistoryCarDto":{"type":"object","properties":{"plateNumber":{"type":"string"}},"required":["plateNumber"]},"UpdateHistoryCarDto":{"type":"object","properties":{"plateNumber":{"type":"string"}}}}}}